New flag -D_USE_EIGER_MODEL removes all computation and inserts call to eiger model.
When compiling an app using lwperf with -D_USE_EIGER_MODEL, the code in between the calls to PERFLOG and PERFSTOP are removed and replaced with a call to SSTMAC_compute_eiger, populated with the appropriate information indicated by the PERFLOG calls. It assumes that models follow the convention of <sitename>.model, where <sitename> is the first parameter to both PERFLOG and PERFSTOP. Use the special macros PERFLOGKEEP and PERFSTOPKEEP when the region of code is to remain inside the skeleton instead of taken out.